Revert "NAS-119008 / 22.12 / properly send HA events on SCALE (#10115)"
This reverts commit 98ec9b5f8b7826ecc9083866510aa3617410e3eb.
