+++ tests/api2/test_014_failover_related.py
@@ -109,28 +109,3 @@ if ha:
             remote = call('failover.call_remote', 'network.configuration.config')
             assert remote['nameserver3'] == old_ns
             assert remote['state']['nameserver3'] == old_ns
-
-def test_08_vrrp_thread_is_running():
-    rv = make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.is_running'})
-    assert isinstance(rv, dict)
-    if not ha:
-        assert not rv['result'], 'VRRP FIFO Thread should not be running on non-HA systems'
-
-        # try to start it
-        make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.start'})
-        rv = make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.is_running'})
-        assert not rv['result'], 'VRRP FIFO Thread should not start on non-HA systems'
-    else:
-        assert rv['result'], 'VRRP FIFO Thread should always be running on HA systems'
-
-        # we stop and start this thread on failover backup events
-
-        # stop it
-        make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.stop'})
-        rv = make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.is_running'})
-        assert not rv['result'], 'VRRP FIFO Thread should stop on HA systems'
-
-        # start it
-        make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.start'})
-        rv = make_ws_request(ip, {'msg': 'method', 'method': 'vrrp.thread.is_running'})
-        assert rv['result'], 'VRRP FIFO Thread should start on HA systems'
